+++ generators/java/generator-utils/src/main/java/com/fern/java/GlobalHeaders.java
@@ -1,8 +1,8 @@
 package com.fern.java;
 
-import com.fern.irV42.model.http.HttpHeader;
-import com.fern.irV42.model.ir.IntermediateRepresentation;
-import com.fern.irV42.model.types.ContainerType;
+import com.fern.ir.model.http.HttpHeader;
+import com.fern.ir.model.ir.IntermediateRepresentation;
+import com.fern.ir.model.types.ContainerType;
 import com.squareup.javapoet.ParameterSpec;
 import java.util.ArrayList;
 import java.util.List;