+++ seed/java-sdk/exhaustive/custom-dependency/src/main/java/com/seed/exhaustive/resources/types/object/types/ObjectWithOptionalField.java
@@ -12,6 +12,7 @@ import com.fasterxml.jackson.annotation.JsonSetter;
 import com.fasterxml.jackson.annotation.Nulls;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.seed.exhaustive.core.ObjectMappers;
+import java.math.BigInteger;
 import java.time.OffsetDateTime;
 import java.util.HashMap;
 import java.util.List;
@@ -48,6 +49,8 @@ public final class ObjectWithOptionalField {
 
     private final Optional<Map<Integer, String>> map;
 
+    private final Optional<BigInteger> bigint;
+
     private final Map<String, Object> additionalProperties;
 
     private ObjectWithOptionalField(
@@ -63,6 +66,7 @@ public final class ObjectWithOptionalField {
             Optional<List<String>> list,
             Optional<Set<String>> set,
             Optional<Map<Integer, String>> map,
+            Optional<BigInteger> bigint,
             Map<String, Object> additionalProperties) {
         this.string = string;
         this.integer = integer;
@@ -76,6 +80,7 @@ public final class ObjectWithOptionalField {
         this.list = list;
         this.set = set;
         this.map = map;
+        this.bigint = bigint;
         this.additionalProperties = additionalProperties;
     }
 
@@ -139,6 +144,11 @@ public final class ObjectWithOptionalField {
         return map;
     }
 
+    @JsonProperty("bigint")
+    public Optional<BigInteger> getBigint() {
+        return bigint;
+    }
+
     @java.lang.Override
     public boolean equals(Object other) {
         if (this == other) return true;
@@ -162,7 +172,8 @@ public final class ObjectWithOptionalField {
                 && base64.equals(other.base64)
                 && list.equals(other.list)
                 && set.equals(other.set)
-                && map.equals(other.map);
+                && map.equals(other.map)
+                && bigint.equals(other.bigint);
     }
 
     @java.lang.Override
@@ -179,7 +190,8 @@ public final class ObjectWithOptionalField {
                 this.base64,
                 this.list,
                 this.set,
-                this.map);
+                this.map,
+                this.bigint);
     }
 
     @java.lang.Override
@@ -217,6 +229,8 @@ public final class ObjectWithOptionalField {
 
         private Optional<Map<Integer, String>> map = Optional.empty();
 
+        private Optional<BigInteger> bigint = Optional.empty();
+
         @JsonAnySetter
         private Map<String, Object> additionalProperties = new HashMap<>();
 
@@ -235,6 +249,7 @@ public final class ObjectWithOptionalField {
             list(other.getList());
             set(other.getSet());
             map(other.getMap());
+            bigint(other.getBigint());
             return this;
         }
 
@@ -370,6 +385,17 @@ public final class ObjectWithOptionalField {
             return this;
         }
 
+        @JsonSetter(value = "bigint", nulls = Nulls.SKIP)
+        public Builder bigint(Optional<BigInteger> bigint) {
+            this.bigint = bigint;
+            return this;
+        }
+
+        public Builder bigint(BigInteger bigint) {
+            this.bigint = Optional.of(bigint);
+            return this;
+        }
+
         public ObjectWithOptionalField build() {
             return new ObjectWithOptionalField(
                     string,
@@ -384,6 +410,7 @@ public final class ObjectWithOptionalField {
                     list,
                     set,
                     map,
+                    bigint,
                     additionalProperties);
         }
     }