+++ seed/ruby-sdk/exhaustive/lib/fern_exhaustive/endpoints/object/client.rb
@@ -33,6 +33,7 @@ module SeedExhaustiveClient
       #   * :list (Array<String>)
       #   * :set (Set<String>)
       #   * :map (Hash{Integer => String})
+      #   * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::ObjectWithOptionalField]
       # @example
@@ -112,6 +113,7 @@ module SeedExhaustiveClient
       #     * :list (Array<String>)
       #     * :set (Set<String>)
       #     * :map (Hash{Integer => String})
+      #     * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::NestedObjectWithOptionalField]
       # @example
@@ -148,6 +150,7 @@ module SeedExhaustiveClient
       #     * :list (Array<String>)
       #     * :set (Set<String>)
       #     * :map (Hash{Integer => String})
+      #     * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::NestedObjectWithRequiredField]
       # @example
@@ -183,6 +186,7 @@ module SeedExhaustiveClient
       #     * :list (Array<String>)
       #     * :set (Set<String>)
       #     * :map (Hash{Integer => String})
+      #     * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::NestedObjectWithRequiredField]
       # @example
@@ -227,6 +231,7 @@ module SeedExhaustiveClient
       #   * :list (Array<String>)
       #   * :set (Set<String>)
       #   * :map (Hash{Integer => String})
+      #   * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::ObjectWithOptionalField]
       # @example
@@ -312,6 +317,7 @@ module SeedExhaustiveClient
       #     * :list (Array<String>)
       #     * :set (Set<String>)
       #     * :map (Hash{Integer => String})
+      #     * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::NestedObjectWithOptionalField]
       # @example
@@ -350,6 +356,7 @@ module SeedExhaustiveClient
       #     * :list (Array<String>)
       #     * :set (Set<String>)
       #     * :map (Hash{Integer => String})
+      #     * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::NestedObjectWithRequiredField]
       # @example
@@ -387,6 +394,7 @@ module SeedExhaustiveClient
       #     * :list (Array<String>)
       #     * :set (Set<String>)
       #     * :map (Hash{Integer => String})
+      #     * :bigint (String)
       # @param request_options [SeedExhaustiveClient::RequestOptions]
       # @return [SeedExhaustiveClient::Types::Object::NestedObjectWithRequiredField]
       # @example