+++ seed/ruby-sdk/exhaustive/lib/fern_exhaustive/inlined_requests/client.rb
@@ -32,6 +32,7 @@ module SeedExhaustiveClient
     #   * :list (Array<String>)
     #   * :set (Set<String>)
     #   * :map (Hash{Integer => String})
+    #   * :bigint (String)
     # @param request_options [SeedExhaustiveClient::RequestOptions]
     # @return [SeedExhaustiveClient::Types::Object::ObjectWithOptionalField]
     # @example
@@ -89,6 +90,7 @@ module SeedExhaustiveClient
     #   * :list (Array<String>)
     #   * :set (Set<String>)
     #   * :map (Hash{Integer => String})
+    #   * :bigint (String)
     # @param request_options [SeedExhaustiveClient::RequestOptions]
     # @return [SeedExhaustiveClient::Types::Object::ObjectWithOptionalField]
     # @example