+++ seed/ts-sdk/exhaustive/dev-dependencies/src/api/resources/endpoints/resources/httpMethods/client/Client.ts
@@ -190,7 +190,8 @@ export class HttpMethods {
      *         set: new Set(["string"]),
      *         map: {
      *             1: "string"
-     *         }
+     *         },
+     *         bigint: "123456789123456789"
      *     })
      */
     public async testPatch(