+++ seed/ts-sdk/exhaustive/no-custom-config/src/api/resources/inlinedRequests/client/requests/PostWithObjectBody.ts
@@ -23,7 +23,8 @@ import * as SeedExhaustive from "../../../../index";
  *             set: new Set(["string"]),
  *             map: {
  *                 1: "string"
- *             }
+ *             },
+ *             bigint: "123456789123456789"
  *         }
  *     }
  */