+++ generators/java/sdk/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - 2024-06-07
+
+- Feature: The generator now supports BigInteger types.
+- Chore: Bump intermediate representation to v46
+
 ## [0.9.8] - 2024-06-06
 
 - Fix: `RequestOptions` are now generated with the `timeout` field initialized to `Optional.empty()` instead of `null`